# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ExamBoost Togo API
  description: |
    Backend API for ExamBoost Togo — Adaptive exam preparation platform for
    Togolese national exams (BEPC, BAC1, BAC2).

    The API exposes:
      - **Authentication** — JWT-based register / login / me endpoints.
      - **Question bank** — Paginated listing with filters (subject, exam,
        serie, competence, chapter), random sampling for exam simulations,
        admin-only creation endpoint.
      - **Learning sessions** — Records each student answer and updates both
        the SM-2 spaced-repetition state and the Bayesian Knowledge Tracing
        (BKT) per-competence mastery estimate.
      - **ML predictions** — Predicted exam score (XGBoost if >= 100
        responses available, heuristic BKT fallback otherwise) and dropout
        risk score.

    Adaptive learning algorithms implemented server-side (mirror of the
    Flutter app's local-first engine):
      - **IRT 3PL** (a/b/c parameters, Fisher information for CAT)
      - **BKT** (4-parameter model: P(L), P(T), P(S), P(G), mastery 0.85)
      - **SM-2** (Easiness Factor >= 1.3, intervals 1/6/N*EF)

    Default BKT parameters (configurable via env vars):
      - P(L) initial  = 0.10
      - P(T) transit  = 0.20
      - P(S) slip     = 0.10
      - P(G) guess    = 0.20
      - mastery       = 0.85
  version: 1.0.0
  contact:
    name: ExamBoost Togo Team
    email: contact@examboost.togo
    url: https://github.com/djabelo712/ExamBoost-Togo
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new student
      description: |
        Creates a new student account and returns a JWT access token
        (valid for 7 days by default). The first account created becomes
        admin automatically (useful for dev bootstrap); in production,
        the env var `ADMIN_EMAIL` overrides this behaviour.
      security: []
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              bepcStudent:
                summary: BEPC student (no serie)
                value:
                  email: "amina.kofi@example.com"
                  password: "monMotDePasse123"
                  nom: "Kofi"
                  prenom: "Amina"
                  niveau_scolaire: "3eme"
              bacStudent:
                summary: Terminale C student
                value:
                  email: "kossi.bla@example.com"
                  password: "monMotDePasse123"
                  nom: "Bla"
                  prenom: "Kossi"
                  niveau_scolaire: "Terminale"
                  serie: "C"
                  etablissement: "Lycée de Tokoin"
                  ville: "Lomé"
      responses:
        '201':
          description: Account created — JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Un compte existe deja avec cet email"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login a student
      description: |
        Authenticates a student with email + password and returns a JWT
        access token (valid for 7 days).
      security: []
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: "amina.kofi@example.com"
              password: "monMotDePasse123"
      responses:
        '200':
          description: Authentication successful — JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Email ou mot de passe incorrect"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current user's profile
      description: |
        Returns the full profile of the authenticated user, including the
        BKT mastery map (`bkt_maitrise` keyed by `competence_id`), the IRT
        theta estimate, and aggregate counters.
      operationId: getCurrentUser
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── Questions ─────────────────────────────────────────────────────
...
  /predict-score/{user_id}:
    get:
      tags: [Prediction]
      summary: Predict the student's exam score
      description: |
        Predicts the score (on 20) the student would get at the given
        exam. Strategy:

          - If `total_responses >= 100` and the XGBoost model is available
            on disk → use **XGBoost** with 8 features (pL_global, pL_maths,
            pL_fr, pL_sciences, sessions_7j, avg_time_per_q,
            simulations_completed, last_sim_score).
          - Otherwise → **heuristic** fallback = average of BKT P(L) per
            subject × 20.

        The response includes a `breakdown` with the estimated score per
        subject, and the `method` used (`heuristic` | `xgboost`).
      security: []
      operationId: predictScore
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
        - name: examen
          in: query
          description: Target exam
          required: false
          schema:
            type: string
            enum: [BEPC, BAC1, BAC2]
            default: BEPC
      responses:
        '200':
          description: Predicted score and breakdown
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictScoreOut'
        '404':
          $ref: '#/components/responses/NotFound'
  /predict-dropout/{user_id}:
    get:
      tags: [Prediction]
      summary: Predict the student's dropout risk
      description: |
        Estimates the probability (0-1) that the student will drop out of
        the app in the next 30 days. Currently a transparent rule-based
        model (for the DJANTA pitch), based on:

          - Inactivity (days since last session)
          - Engagement (sessions in the last 7 days)
          - Performance (global BKT mastery)
          - Average time spent per question
          - Number of completed simulations

        Will be replaced by an XGBoost classifier once enough labeled data
        is collected.
      security: []
      operationId: predictDropout
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dropout risk prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictDropoutOut'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/auth/register` or `/auth/login`.
        Pass it as `Authorization: Bearer <token>` in the request header.
        Tokens are valid for 7 days (configurable via
        `ACCESS_TOKEN_EXPIRE_MINUTES`).
  schemas:
    UserCreate:
      type: object
      description: Payload for student registration.
      required: [email, password, nom, prenom, niveau_scolaire]
      properties:
        email:
          type: string
          format: email
          example: "amina.kofi@example.com"
        password:
          type: string
          minLength: 6
          maxLength: 128
          description: Plain-text password (bcrypt-hashed server-side)
          example: "monMotDePasse123"
        nom:
          type: string
          minLength: 1
          maxLength: 80
          example: "Kofi"
        prenom:
          type: string
          minLength: 1
          maxLength: 80
          example: "Amina"
        niveau_scolaire:
          type: string
          description: Grade level
          enum: [3eme, 2nde, 1ere, Terminale, tle]
          example: "3eme"
        serie:
          type: string
          nullable: true
          description: Track (required only for 1ère and Terminale). Null for BEPC.
          enum: [A, B, C, D, F]
          example: "C"
        etablissement:
          type: string
          nullable: true
          example: "Lycée de Tokoin"
        ville:
          type: string
          nullable: true
          example: "Lomé"
    UserLogin:
      type: object
      description: Payload for student login.
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserOut:
      type: object
      description: Public user profile (no password hash).
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
        nom:
          type: string
        prenom:
          type: string
        niveau_scolaire:
          type: string
          enum: [3eme, 2nde, 1ere, Terminale]
        serie:
          type: string
          nullable: true
          enum: [A, B, C, D, F]
        etablissement:
          type: string
          nullable: true
        ville:
          type: string
          nullable: true
        date_inscription:
          type: string
          format: date-time
        theta_irt:
          type: number
          nullable: true
          description: IRT 3PL ability estimate (theta). Null until calibrated.
        total_sessions:
          type: integer
          minimum: 0
        total_questions_answered:
          type: integer
          minimum: 0
        bkt_maitrise:
          type: object
          description: |
            Map of competence_id → BKT P(L) (probability the student has
            mastered the competence). Values are floats in [0, 1].
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
          example:
            "TG-MATHS-ALGEBRE-EQ2D": 0.4667
            "TG-MATHS-GEOMETRIE-PYTHAGORE": 0.8234
    Token:
      type: object
      description: JWT authentication response.
      properties:
        access_token:
          type: string
          description: JWT token (HS256 signed)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user_id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserOut'
...
    ScoreBreakdownItem:
      type: object
      description: Estimated score for one subject.
      properties:
        matiere:
          type: string
          example: "Mathematiques"
        score_estime:
          type: number
          minimum: 0
          maximum: 20
          description: Estimated score on 20 (= pL_moyen × 20)
          example: 9.33
        pL_moyen:
          type: number
          minimum: 0
          maximum: 1
          description: Average BKT P(L) for the subject
          example: 0.4667
        nb_questions:
          type: integer
          minimum: 0
          description: Number of questions answered in this subject
          example: 8
    PredictScoreOut:
      type: object
      description: Predicted exam score (on 20) with breakdown by subject.
      properties:
        user_id:
          type: string
          format: uuid
        examen:
          type: string
          enum: [BEPC, BAC1, BAC2]
        predicted_score:
          type: number
          minimum: 0
          maximum: 20
          example: 9.33
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: |
            Model confidence. Low (< 0.5) for heuristic method or when few
            responses have been recorded; high (> 0.7) for XGBoost with
            sufficient data.
          example: 0.3
        method:
          type: string
          enum: [heuristic, xgboost, insufficient_data]
          description: Prediction method used
          example: "heuristic"
        breakdown:
          type: array
          description: Estimated score per subject
          items:
            $ref: '#/components/schemas/ScoreBreakdownItem'
        total_responses:
          type: integer
          minimum: 0
          description: Total number of responses recorded for this user
          example: 12
    PredictDropoutOut:
      type: object
      description: Dropout risk prediction.
      properties:
        user_id:
          type: string
          format: uuid
        dropout_probability:
          type: number
          minimum: 0
          maximum: 1
          description: Probability of dropping out in the next 30 days
          example: 0.42
        risk_level:
          type: string
          enum: [faible, modere, eleve]
          description: Categorical risk level (low / moderate / high)
          example: "modere"
        factors:
          type: object
          description: |
            Key factors driving the prediction (transparent model). Keys are
            factor names, values are normalised contributions.
          additionalProperties: true
          example:
            inactivity_days: 5
            sessions_7j: 2
            pL_global: 0.42
            avg_time_per_q: 38.5
            simulations_completed: 1
    Error:
      type: object
      description: Standard error response (FastAPI format).
      properties:
        detail:
          oneOf:
            - type: string
              example: "Utilisateur 550e8400-e29b-41d4-a716-446655440000 introuvable"
            - type: object
              description: Validation error detail (FastAPI wraps multiple errors)
            - type: array
              items: { type: object }
  responses:
    BadRequest:
      description: Bad request — validation error or malformed payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "niveau_scolaire doit etre dans ['2nde', '3eme', '1ere', 'Terminale', 'tle']"
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Token invalide ou absent"
    NotFound:
      description: Not found — resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Question TG-BEPC-MATHS-2022-Q99 introuvable"
...